[feat]add child routes for servers and users
previously, routes where only using 1 router outlet which resultedin routes that where a sub-route replacing the content of the parent route entirely. To resolve this, we made sub-routes children of the parent route in the routing module and added nested router outlets for parent routes that needed to display child content dynamically.